server-api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-cloud-aiplatform==1.42.1
google-auth==2.28.2
requests==2.31.0
python-dotenv